% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_crosscut.R
\name{crosscut_assemble}
\alias{crosscut_assemble}
\title{Assemble a "resampled" bullet sequence}
\usage{
crosscut_assemble(len, tab = c("bullet.slice", "bullet.slice.idx"),
  con = NULL)
}
\arguments{
\item{len}{target sequence length}

\item{tab}{if a db connection is passed in, this should be a character vector
of length two, giving the name of the full table and the summary table.
The full table should have (at least) columns id, sig, and type. The
summary table should have columns id, type, n (the length of the chunk),
and not_na (the number of non-na pieces). The full table should have all
sequence pieces which are to be resampled. If a scalar is supplied, the
summary table will be generated within the function.
If a db connection is not available, this should be a data frame with the
relevant information.}

\item{con}{database connection}
}
\description{
Assemble a "resampled" bullet sequence
}
