% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_crosscut.R
\name{cycle_draw}
\alias{cycle_draw}
\title{Draw cycles for a resampled bullet sequence}
\usage{
cycle_draw(len, tab, boundary_indicator = "boundary")
}
\arguments{
\item{len}{target sequence length}

\item{tab}{summary table}

\item{boundary_indicator}{character indicating an incomplete cycle}
}
\value{
a data frame of sampled responses, in order
}
\description{
This function assembles cycles to form a sequence of approximately length
`len` using a summary table with the following information:
\enumerate{
\item id, a variable uniquely identifying the cycle,
\item type, a character or factor variable with values `boundary_indicator`
        and some other value indicating a complete cycle
\item n, the length of the sequence,
\item not_na, the number of non NA values in the sequence
}
}
