% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_crosscut.R
\name{df_fix_length}
\alias{df_fix_length}
\title{Pad or trim the data frame to the specified number of rows}
\usage{
df_fix_length(full_df, len)
}
\arguments{
\item{full_df}{full data frame to pad/trim to the correct length}

\item{len}{desired length of the data frame}
}
\value{
a data frame with len rows
}
\description{
Pad or trim the data frame to the specified number of rows
}
