% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sig}
\alias{sig}
\title{Identified signature of Barrel 1, Bullet 1, Land 3 from the Hamby 252 study}
\format{a data frame with columns x, y, value, raw_sig, se, and sig. All are
          numeric.}
\usage{
sig
}
\description{
Identified signature of Barrel 1, Bullet 1, Land 3 from the Hamby 252 study
}
\keyword{datasets}
