% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{signature_slices}
\alias{signature_slices}
\title{Cycle slices from Hamby252 Barrel 1, Bullets 1 and 2}
\format{a data frame with columns id, type, x, y, value, sig. id and type are
          character, x, y, value, and sig are numeric}
\usage{
signature_slices
}
\description{
Cycle slices from Hamby252 Barrel 1, Bullets 1 and 2
}
\keyword{datasets}
