% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cycle.R
\name{crosscut_slice}
\alias{crosscut_slice}
\title{Slice signature into cycles}
\usage{
crosscut_slice(x)
}
\arguments{
\item{x}{a numeric vector with signature information}
}
\value{
a list of vectors with signature cycles
}
\description{
Slice signature into cycles
}
\examples{
data("sig")
crosscut_slice(sig) # Slice whole data frame
crosscut_slice(sig$sig) # Slice signature only
}
